+++ Cargo.toml
@@ -6,12 +6,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bstr = { version = "1.1.0", default-features = false, features = ["std"] }
+bstr = "1.1.0"
 bytecount = "0.6"
 clap = { version = "4.3.0", features = ["derive"] }
 encoding_rs = "0.8.14"
 encoding_rs_io = "0.1.6"
-grep = { git = "https://github.com/helixbass/ripgrep", rev = "b83cc3c" }
 ignore = "0.4.20"
 libc = "0.2.144"
 libloading = "0.8.0"
