+++ src/matcher/mod.rs
@@ -225,173 +225,3 @@ pub enum LineMatchKind {
     Confirmed(usize),
     Candidate(usize),
 }
-
-pub trait Matcher {
-    type Error: fmt::Display;
-
-    fn find_at(&self, haystack: &[u8], at: usize) -> Result<Option<Match>, Self::Error>;
-
-    fn find(&self, haystack: &[u8]) -> Result<Option<Match>, Self::Error> {
-        self.find_at(haystack, 0)
-    }
-
-    fn find_iter<F>(&self, haystack: &[u8], matched: F) -> Result<(), Self::Error>
-    where
-        F: FnMut(Match) -> bool,
-    {
-        self.find_iter_at(haystack, 0, matched)
-    }
-
-    fn find_iter_at<F>(&self, haystack: &[u8], at: usize, mut matched: F) -> Result<(), Self::Error>
-    where
-        F: FnMut(Match) -> bool,
-    {
-        self.try_find_iter_at(haystack, at, |m| Ok(matched(m)))
-            .map(|r: Result<(), ()>| r.unwrap())
-    }
-
-    fn try_find_iter<F, E>(&self, haystack: &[u8], matched: F) -> Result<Result<(), E>, Self::Error>
-    where
-        F: FnMut(Match) -> Result<bool, E>,
-    {
-        self.try_find_iter_at(haystack, 0, matched)
-    }
-
-    fn try_find_iter_at<F, E>(
-        &self,
-        haystack: &[u8],
-        at: usize,
-        mut matched: F,
-    ) -> Result<Result<(), E>, Self::Error>
-    where
-        F: FnMut(Match) -> Result<bool, E>,
-    {
-        let mut last_end = at;
-        let mut last_match = None;
-
-        loop {
-            if last_end > haystack.len() {
-                return Ok(Ok(()));
-            }
-            let m = match self.find_at(haystack, last_end)? {
-                None => return Ok(Ok(())),
-                Some(m) => m,
-            };
-            if m.start == m.end {
-                last_end = m.end + 1;
-                if Some(m.end) == last_match {
-                    continue;
-                }
-            } else {
-                last_end = m.end;
-            }
-            last_match = Some(m.end);
-            match matched(m) {
-                Ok(true) => continue,
-                Ok(false) => return Ok(Ok(())),
-                Err(err) => return Ok(Err(err)),
-            }
-        }
-    }
-
-    fn is_match(&self, haystack: &[u8]) -> Result<bool, Self::Error> {
-        self.is_match_at(haystack, 0)
-    }
-
-    fn is_match_at(&self, haystack: &[u8], at: usize) -> Result<bool, Self::Error> {
-        Ok(self.shortest_match_at(haystack, at)?.is_some())
-    }
-
-    fn shortest_match(&self, haystack: &[u8]) -> Result<Option<usize>, Self::Error> {
-        self.shortest_match_at(haystack, 0)
-    }
-
-    fn shortest_match_at(&self, haystack: &[u8], at: usize) -> Result<Option<usize>, Self::Error> {
-        Ok(self.find_at(haystack, at)?.map(|m| m.end))
-    }
-
-    fn non_matching_bytes(&self) -> Option<&ByteSet> {
-        None
-    }
-
-    fn line_terminator(&self) -> Option<LineTerminator> {
-        None
-    }
-
-    fn find_candidate_line(&self, haystack: &[u8]) -> Result<Option<LineMatchKind>, Self::Error> {
-        Ok(self.shortest_match(haystack)?.map(LineMatchKind::Confirmed))
-    }
-}
-
-impl<'a, M: Matcher> Matcher for &'a M {
-    type Error = M::Error;
-
-    fn find_at(&self, haystack: &[u8], at: usize) -> Result<Option<Match>, Self::Error> {
-        (*self).find_at(haystack, at)
-    }
-
-    fn find(&self, haystack: &[u8]) -> Result<Option<Match>, Self::Error> {
-        (*self).find(haystack)
-    }
-
-    fn find_iter<F>(&self, haystack: &[u8], matched: F) -> Result<(), Self::Error>
-    where
-        F: FnMut(Match) -> bool,
-    {
-        (*self).find_iter(haystack, matched)
-    }
-
-    fn find_iter_at<F>(&self, haystack: &[u8], at: usize, matched: F) -> Result<(), Self::Error>
-    where
-        F: FnMut(Match) -> bool,
-    {
-        (*self).find_iter_at(haystack, at, matched)
-    }
-
-    fn try_find_iter<F, E>(&self, haystack: &[u8], matched: F) -> Result<Result<(), E>, Self::Error>
-    where
-        F: FnMut(Match) -> Result<bool, E>,
-    {
-        (*self).try_find_iter(haystack, matched)
-    }
-
-    fn try_find_iter_at<F, E>(
-        &self,
-        haystack: &[u8],
-        at: usize,
-        matched: F,
-    ) -> Result<Result<(), E>, Self::Error>
-    where
-        F: FnMut(Match) -> Result<bool, E>,
-    {
-        (*self).try_find_iter_at(haystack, at, matched)
-    }
-
-    fn is_match(&self, haystack: &[u8]) -> Result<bool, Self::Error> {
-        (*self).is_match(haystack)
-    }
-
-    fn is_match_at(&self, haystack: &[u8], at: usize) -> Result<bool, Self::Error> {
-        (*self).is_match_at(haystack, at)
-    }
-
-    fn shortest_match(&self, haystack: &[u8]) -> Result<Option<usize>, Self::Error> {
-        (*self).shortest_match(haystack)
-    }
-
-    fn shortest_match_at(&self, haystack: &[u8], at: usize) -> Result<Option<usize>, Self::Error> {
-        (*self).shortest_match_at(haystack, at)
-    }
-
-    fn non_matching_bytes(&self) -> Option<&ByteSet> {
-        (*self).non_matching_bytes()
-    }
-
-    fn line_terminator(&self) -> Option<LineTerminator> {
-        (*self).line_terminator()
-    }
-
-    fn find_candidate_line(&self, haystack: &[u8]) -> Result<Option<LineMatchKind>, Self::Error> {
-        (*self).find_candidate_line(haystack)
-    }
-}
